@@ -91,11 +91,8 @@ class Finkok
         ],
     ];
 
-    private FinkokSettings $settings;
-
-    public function __construct(FinkokSettings $factory)
+    public function __construct(private FinkokSettings $settings)
     {
-        $this->settings = $factory;
     }
 
     public function settings(): FinkokSettings
@@ -123,17 +120,21 @@ class Finkok
      * @param mixed $command
      * @return object|null
      */
-    protected function checkCommand(string $method, $command): ?object
+    protected function checkCommand(string $method, mixed $command): ?object
     {
         $expected = static::SERVICES_MAP[$method][1];
         if ('' === $expected) {
             return null;
         }
         if (! is_object($command) || ! is_a($command, $expected)) {
-            $type = (is_object($command)) ? get_class($command) : gettype($command);
-            throw new InvalidArgumentException(
-                sprintf('Call %s::%s expect %s but received %s', static::class, $method, $expected, $type)
+            $message = sprintf(
+                'Call %s::%s expect %s but received %s',
+                static::class,
+                $method,
+                $expected,
+                get_debug_type($command),
             );
+            throw new InvalidArgumentException($message);
         }
         return $command;
     }
@@ -159,7 +160,7 @@ class Finkok
         $method = static::SERVICES_MAP[$method][2] ?? $method;
         if (! is_callable([$service, $method])) {
             throw new BadMethodCallException(
-                sprintf('The service %s does not have a method %s', get_class($service), $method)
+                sprintf('The service %s does not have a method %s', $service::class, $method)
             );
         }
         return $service->{$method}($command);
