@@ -12,8 +12,6 @@ use PhpCfdi\XmlCancelacion\XmlCancelacionHelper;
 
 class CancelSigner
 {
-    private CancelDocuments $documents;
-
     private DateTimeImmutable $dateTime;
 
     /**
@@ -22,9 +20,8 @@ class CancelSigner
      * @param CancelDocuments $documents
      * @param DateTimeImmutable|null $dateTime If null or ommited then use current time and time zone
      */
-    public function __construct(CancelDocuments $documents, ?DateTimeImmutable $dateTime = null)
+    public function __construct(private CancelDocuments $documents, ?DateTimeImmutable $dateTime = null)
     {
-        $this->documents = $documents;
         $this->dateTime = $dateTime ?? new DateTimeImmutable();
     }
 
