@@ -14,8 +14,6 @@ class GetRelatedSigner
     /** @var string */
     public const DEFAULT_PACRFC = 'CVD110412TF6';
 
-    private string $uuid;
-
     private RfcRole $role;
 
     private string $pacRfc;
@@ -27,9 +25,8 @@ class GetRelatedSigner
      * @param RfcRole|null $role If null or omitted then uses issuer role
      * @param string $pacRfc If empty or omitted then uses DEFAULT_PACRFC
      */
-    public function __construct(string $uuid, RfcRole $role = null, string $pacRfc = self::DEFAULT_PACRFC)
+    public function __construct(private string $uuid, ?RfcRole $role = null, string $pacRfc = self::DEFAULT_PACRFC)
     {
-        $this->uuid = $uuid;
         $this->role = $role ?? RfcRole::issuer();
         $this->pacRfc = $pacRfc ?: static::DEFAULT_PACRFC;
     }
