@@ -8,17 +8,11 @@ use PhpCfdi\Finkok\Definitions\CancelAnswer;
 
 class AcceptRejectUuidItem
 {
-    private string $uuid;
-
-    private AcceptRejectUuidStatus $status;
-
-    private CancelAnswer $answer;
-
-    public function __construct(string $uuid, AcceptRejectUuidStatus $status, CancelAnswer $answer)
-    {
-        $this->uuid = $uuid;
-        $this->status = $status;
-        $this->answer = $answer;
+    public function __construct(
+        private string $uuid,
+        private AcceptRejectUuidStatus $status,
+        private CancelAnswer $answer,
+    ) {
     }
 
     public function uuid(): string
