@@ -8,8 +8,6 @@ use PhpCfdi\Finkok\Definitions\CancelStorePending;
 
 class CancelSignatureCommand
 {
-    private string $xml;
-
     private CancelStorePending $storePending;
 
     /**
@@ -18,9 +16,8 @@ class CancelSignatureCommand
      * @param string $xml The signed xml
      * @param CancelStorePending|null $storePending Defaults to CancelStorePending::no()
      */
-    public function __construct(string $xml, CancelStorePending $storePending = null)
+    public function __construct(private string $xml, ?CancelStorePending $storePending = null)
     {
-        $this->xml = $xml;
         $this->storePending = $storePending ?? CancelStorePending::no();
     }
 
