@@ -6,11 +6,8 @@ namespace PhpCfdi\Finkok\Services\Cancel;
 
 class GetRelatedSignatureCommand
 {
-    private string $xml;
-
-    public function __construct(string $xml)
+    public function __construct(private string $xml)
     {
-        $this->xml = $xml;
     }
 
     public function xml(): string
