@@ -6,17 +6,8 @@ namespace PhpCfdi\Finkok\Services\Cancel;
 
 class RelatedItem
 {
-    private string $uuid;
-
-    private string $rfcEmitter;
-
-    private string $rfcReceiver;
-
-    public function __construct(string $uuid, string $rfcEmitter, string $rfcReceiver)
+    public function __construct(private string $uuid, private string $rfcEmitter, private string $rfcReceiver)
     {
-        $this->uuid = $uuid;
-        $this->rfcEmitter = $rfcEmitter;
-        $this->rfcReceiver = $rfcReceiver;
     }
 
     public function uuid(): string
