@@ -6,28 +6,16 @@ namespace PhpCfdi\Finkok\Services\Registration;
 
 class AddCommand
 {
-    private string $rfc;
-
     private CustomerType $type;
 
-    private string $certificate;
-
-    private string $privateKey;
-
-    private string $passPhrase;
-
     public function __construct(
-        string $rfc,
+        private string $rfc,
         ?CustomerType $type = null,
-        string $certificate = '',
-        string $privateKey = '',
-        string $passPhrase = ''
+        private string $certificate = '',
+        private string $privateKey = '',
+        private string $passPhrase = ''
     ) {
-        $this->rfc = $rfc;
         $this->type = $type ?? CustomerType::ondemand();
-        $this->certificate = $certificate;
-        $this->privateKey = $privateKey;
-        $this->passPhrase = $passPhrase;
     }
 
     public function rfc(): string
