@@ -6,11 +6,8 @@ namespace PhpCfdi\Finkok\Services\Stamping;
 
 class QueryPendingCommand
 {
-    private string $uuid;
-
-    public function __construct(string $uuid)
+    public function __construct(private string $uuid)
     {
-        $this->uuid = $uuid;
     }
 
     public function uuid(): string
