@@ -13,7 +13,7 @@ class SoapFactory implements LoggerAwareInterface
 {
     private LoggerInterface $logger;
 
-    public function __construct(LoggerInterface $logger = null)
+    public function __construct(?LoggerInterface $logger = null)
     {
         $this->logger = $logger ?? new NullLogger();
     }
