@@ -18,7 +18,7 @@ final class FakeSoapFactory extends SoapFactory
 
     public stdClass $preparedResult;
 
-    public function __construct(LoggerInterface $logger = null)
+    public function __construct(?LoggerInterface $logger = null)
     {
         parent::__construct($logger);
         $this->preparedResult = (object) [];
